refactor: crates
* split overlay crate into overlay, dll, vulkan-layer
* move ipc server, client impl into overlay, client crate
* split event crate from common
