+++ crates/client/Cargo.toml
@@ -10,8 +10,11 @@ readme = "../../README.md"
 
 [dependencies]
 asdf-overlay-common = { workspace = true }
-tokio = { workspace = true, features = ["macros", "time"] }
+asdf-overlay-event = { workspace = true, features = ["bincode"] }
+tokio = { workspace = true, features = ["macros", "time", "net"] }
 
+bincode = "2.0.1"
+dashmap = "6.1.0"
 ntapi = "0.4.1"
 anyhow = "1.0.97"
 scopeguard = "1.2.0"
