+++ crates/common/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod cursor;
-pub mod event;
 pub mod ipc;
-pub mod key;
 pub mod request;
 pub mod size;
