+++ crates/overlay/Cargo.toml
@@ -8,37 +8,26 @@ repository.workspace = true
 authors = ["storycraft <storycraft@pancake.sh>"]
 readme = "../../README.md"
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
 asdf-overlay-common = { workspace = true }
+asdf-overlay-event = { workspace = true }
 asdf-overlay-hook = { workspace = true }
-tokio = { workspace = true, features = ["full"] }
 
+tracing = "0.1"
 anyhow = "1.0.97"
 parking_lot = { version = "0.12.3", features = ["hardware-lock-elision"] }
-serde = "1.0.219"
-tokio-util = "0.7.14"
-futures = "0.3.31"
 once_cell = "1.21.3"
 scopeguard = "1.2.0"
 dashmap = "6.1.0"
 nohash-hasher = "0.2.0"
-tracing-subscriber = "0.3"
 ash = { version = "0.38.0", default-features = false, features = ["std"] }
 array-util = "1.0.2"
 windows-link = "0.1.1"
 bytemuck = "1.22.0"
-bitvec = "1"
 utf16string = "0.2.0"
 bitflags = "2.9.1"
 arc-swap = "1.7.1"
 sync_wrapper = "1.0.2"
-
-[dependencies.tracing]
-version = "0.1"
-features = ["max_level_trace", "release_max_level_off"]
 
 [dependencies.windows]
 workspace = true
@@ -62,7 +51,6 @@ features = [
     "Win32_Graphics_Direct3D11",
     "Win32_Graphics_Direct3D12",
     "Win32_Security",
-    "Win32_System_Diagnostics_Debug",
     "Win32_UI_Input_KeyboardAndMouse",
     "Win32_UI_Input_Ime",
     "Win32_Globalization"
