+++ crates/overlay/src/backend/window/cursor.rs
@@ -8,13 +8,13 @@ use windows::{
     core::PCWSTR,
 };
 
-use crate::instance;
-
-use crate::resources::cursors::{
-    IDC_ALIAS, IDC_CELL, IDC_COLRESIZE, IDC_COPYCUR, IDC_HAND_GRAB, IDC_HAND_GRABBING,
-    IDC_PAN_EAST, IDC_PAN_MIDDLE, IDC_PAN_MIDDLE_HORIZONTAL, IDC_PAN_MIDDLE_VERTICAL,
-    IDC_PAN_NORTH, IDC_PAN_NORTH_EAST, IDC_PAN_NORTH_WEST, IDC_PAN_SOUTH, IDC_PAN_SOUTH_EAST,
-    IDC_PAN_SOUTH_WEST, IDC_PAN_WEST, IDC_ROWRESIZE, IDC_VERTICALTEXT, IDC_ZOOMIN, IDC_ZOOMOUT,
+use crate::{
+    instance, resources::cursors::{
+        IDC_ALIAS, IDC_CELL, IDC_COLRESIZE, IDC_COPYCUR, IDC_HAND_GRAB, IDC_HAND_GRABBING,
+        IDC_PAN_EAST, IDC_PAN_MIDDLE, IDC_PAN_MIDDLE_HORIZONTAL, IDC_PAN_MIDDLE_VERTICAL,
+        IDC_PAN_NORTH, IDC_PAN_NORTH_EAST, IDC_PAN_NORTH_WEST, IDC_PAN_SOUTH, IDC_PAN_SOUTH_EAST,
+        IDC_PAN_SOUTH_WEST, IDC_PAN_WEST, IDC_ROWRESIZE, IDC_VERTICALTEXT, IDC_ZOOMIN, IDC_ZOOMOUT,
+    }
 };
 
 pub fn load_cursor(cursor: Cursor) -> Option<HCURSOR> {
