+++ crates/vulkan-layer/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "asdf-overlay-vulkan-layer"
+description = "Asdf Overlay vulkan layer"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors = ["storycraft <storycraft@pancake.sh>"]
+readme = "../../README.md"
+
+[dependencies]
+asdf-overlay = { workspace = true }
+
+tracing = "0.1"
+once_cell = "1.21.3"
+dashmap = "6.1.0"
+nohash-hasher = "0.2.0"
+ash = { version = "0.38.0", default-features = false, features = ["std"] }
