+++ crates/vulkan-layer/src/map.rs
@@ -1,0 +1,4 @@
+use dashmap::DashMap;
+use nohash_hasher::BuildNoHashHasher;
+
+pub type IntDashMap<K, V> = DashMap<K, V, BuildNoHashHasher<K>>;
