+++ xtask/Cargo.toml
@@ -8,5 +8,5 @@ authors = ["storycraft <storycraft@pancake.sh>"]
 [dependencies]
 anyhow = "1.0.97"
 camino = "1.1.9"
-cargo_metadata = "0.19.2"
+cargo_metadata = "0.21.0"
 clap = { version = "4.5.35", features = ["derive"] }
